// angularapp/src/app/model/feedback/feedback-module.ts
import { User } from '../user/user-module';

export interface Feedback {
  id?: number;       // optional because it's auto-generated by backend
  message: string;   // feedback text
  rating?: number;   // rating (1 to 5 stars) - optional
  user?: User;       // optional: reference to User who gave feedback
  userName?: string; // optional: user name for display
  createdAt?: string; // optional: creation timestamp
}